Fix conversation rename and export on new conversation; Clean up refs;
