+++ app/settings/training/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import Training from '@/components/interactive/Settings/training';
-import useUser from '@/components/jrg/auth/hooks/useUser';
 import { SidebarPage } from '@/components/jrg/appwrapper/SidebarPage';
+import { useUser } from '@/components/jrg/auth/hooks/useUser';
 export default function TrainingPage() {
   const { data: user, isLoading } = useUser();
