+++ components/interactive/Settings/training.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useCompany } from '@/components/jrg/auth/hooks/useTeam';
+import { useCompany } from '@/components/jrg/auth/hooks/useUser';
 import { Alert, AlertDescription } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
