+++ components/interactive/hooks/useAgent.ts
@@ -3,10 +3,10 @@ import { getCookie, setCookie } from 'cookies-next';
 import { useContext } from 'react';
 import useSWR, { SWRResponse } from 'swr';
 import { z } from 'zod';
-import { useCompanies } from '../../jrg/auth/hooks/useTeam';
+import { useCompanies } from '../../jrg/auth/hooks/useUser';
 import log from '../../jrg/next-log/log';
 import { InteractiveConfigContext } from '../InteractiveConfigContext';
-import { chainMutations } from './lib';
+import { chainMutations, createGraphQLClient } from './lib';
 
 export const AgentSchema = z.object({
   companyId: z.string().uuid(),
