+++ components/jrg/appwrapper/AppWrapperCenterAlignedBox.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react';
+import { ReactNode } from 'react';
 
 export type CenterAlignedBoxProps = {
   left?: ReactNode;
