+++ components/jrg/appwrapper/Head.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function Head() {
   return (
     <head>
