+++ components/jrg/auth/hooks/useInvitation.ts
@@ -1,11 +1,19 @@
 import { createGraphQLClient } from '@/components/interactive/hooks/lib';
 import useSWR, { SWRResponse } from 'swr';
+import { z } from 'zod';
 import log from '../../next-log/log';
-import { Invitation, InvitationSchema } from '../types/types';
-// ============================================================================
-// Invitation Related Hooks
-// ============================================================================
 
+export const InvitationSchema = z.object({
+  id: z.string().uuid(),
+  companyId: z.string().uuid(),
+  email: z.string().email(),
+  createdAt: z.string().datetime(),
+  inviterId: z.string().uuid(),
+  isAccepted: z.boolean(),
+  roleId: z.string().uuid(),
+});
+
+export type Invitation = z.infer<typeof InvitationSchema>;
 /**
  * Hook to fetch and manage invitations
  * @param companyId - Optional company ID to fetch invitations for
